Reuse more existing Reva code
