--- pkg/sdk/common/crypto/adler32.go
@@ -1,34 +1,0 @@
-// Copyright 2018-2020 CERN
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//     http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-//
-// In applying this license, CERN does not waive the privileges and immunities
-// granted to it by virtue of its status as an Intergovernmental Organization
-// or submit itself to any jurisdiction.
-
-package crypto
-
-import (
-	"fmt"
-	"hash/adler32"
-	"io"
-)
-
-// ComputeAdler32Checksum calculates an Adler32 checksum.
-func ComputeAdler32Checksum(data io.Reader) (string, error) {
-	h := adler32.New()
-	if _, err := io.Copy(h, data); err != nil {
-		return "", err
-	}
-	return fmt.Sprintf("%x", h.Sum(nil)), nil
-}
